refactor: centralize useWorker hook
